Moved traj_to_gif and povray_render files to utils folder Used python_env_check() to update the analyse_forces, so it fixes the FutureWarning Error
